+++ src/fides/api/ops/api/v1/endpoints/privacy_request_endpoints.py
@@ -32,7 +32,6 @@ from starlette.status import (
 
 from fides.api.ops import common_exceptions
 from fides.api.ops.api import deps
-from fides.api.ops.api.v1 import scope_registry as scopes
 from fides.api.ops.api.v1.endpoints.dataset_endpoints import _get_connection_config
 from fides.api.ops.api.v1.endpoints.manual_webhook_endpoints import (
     get_access_manual_webhook_or_404,
@@ -40,6 +39,8 @@ from fides.api.ops.api.v1.endpoints.manual_webhook_endpoints import (
 from fides.api.ops.api.v1.scope_registry import (
     PRIVACY_REQUEST_CALLBACK_RESUME,
     PRIVACY_REQUEST_CREATE,
+    PRIVACY_REQUEST_NOTIFICATIONS_CREATE_OR_UPDATE,
+    PRIVACY_REQUEST_NOTIFICATIONS_READ,
     PRIVACY_REQUEST_READ,
     PRIVACY_REQUEST_REVIEW,
     PRIVACY_REQUEST_UPLOAD_DATA,
@@ -53,6 +54,7 @@ from fides.api.ops.api.v1.urn_registry import (
     PRIVACY_REQUEST_DENY,
     PRIVACY_REQUEST_MANUAL_ERASURE,
     PRIVACY_REQUEST_MANUAL_INPUT,
+    PRIVACY_REQUEST_NOTIFICATIONS,
     PRIVACY_REQUEST_RESUME,
     PRIVACY_REQUEST_RESUME_FROM_REQUIRES_INPUT,
     PRIVACY_REQUEST_RETRY,
@@ -89,6 +91,7 @@ from fides.api.ops.models.privacy_request import (
     CheckpointActionRequired,
     ExecutionLog,
     PrivacyRequest,
+    PrivacyRequestNotifications,
     PrivacyRequestStatus,
     ProvidedIdentity,
     ProvidedIdentityType,
@@ -111,6 +114,7 @@ from fides.api.ops.schemas.privacy_request import (
     ExecutionLogDetailResponse,
     ManualWebhookData,
     PrivacyRequestCreate,
+    PrivacyRequestNotificationInfo,
     PrivacyRequestResponse,
     PrivacyRequestVerboseResponse,
     ReviewPrivacyRequestIds,
@@ -120,6 +124,7 @@ from fides.api.ops.schemas.privacy_request import (
 from fides.api.ops.schemas.redis_cache import Identity
 from fides.api.ops.service._verification import send_verification_code_to_user
 from fides.api.ops.service.messaging.message_dispatch_service import (
+    check_and_dispatch_error_notifications,
     dispatch_message_task,
 )
 from fides.api.ops.service.privacy_request.request_runner_service import (
@@ -513,7 +518,7 @@ def attach_resume_instructions(privacy_request: PrivacyRequest) -> None:
 
 @router.get(
     PRIVACY_REQUESTS,
-    dependencies=[Security(verify_oauth_client, scopes=[scopes.PRIVACY_REQUEST_READ])],
+    dependencies=[Security(verify_oauth_client, scopes=[PRIVACY_REQUEST_READ])],
     response_model=Page[
         Union[
             PrivacyRequestVerboseResponse,
@@ -605,7 +610,7 @@ def get_request_status(
 
 @router.get(
     REQUEST_STATUS_LOGS,
-    dependencies=[Security(verify_oauth_client, scopes=[scopes.PRIVACY_REQUEST_READ])],
+    dependencies=[Security(verify_oauth_client, scopes=[PRIVACY_REQUEST_READ])],
     response_model=Page[ExecutionLogDetailResponse],
 )
 def get_request_status_logs(
@@ -629,6 +634,69 @@ def get_request_status_logs(
         .filter(ExecutionLog.privacy_request_id == privacy_request_id)
         .order_by(ExecutionLog.updated_at.asc()),
         params,
+    )
+
+
+@router.get(
+    PRIVACY_REQUEST_NOTIFICATIONS,
+    status_code=HTTP_200_OK,
+    response_model=PrivacyRequestNotificationInfo,
+    dependencies=[
+        Security(
+            verify_oauth_client,
+            scopes=[PRIVACY_REQUEST_NOTIFICATIONS_READ],
+        )
+    ],
+)
+def get_privacy_request_notification_info(
+    *, db: Session = Depends(deps.get_db)
+) -> PrivacyRequestNotificationInfo:
+    """Retrieve privacy request notification email addresses and number of failures to trigger notifications."""
+    info = PrivacyRequestNotifications.all(db)
+
+    if not info:
+        raise HTTPException(
+            status_code=HTTP_404_NOT_FOUND,
+            detail="No privacy request notification info found",
+        )
+
+    return PrivacyRequestNotificationInfo(
+        email_addresses=[x for x in info[0].email.split(", ")],
+        notify_after_failures=info[0].notify_after_failures,
+    )
+
+
+@router.put(
+    PRIVACY_REQUEST_NOTIFICATIONS,
+    status_code=HTTP_200_OK,
+    response_model=PrivacyRequestNotificationInfo,
+    dependencies=[
+        Security(
+            verify_oauth_client,
+            scopes=[PRIVACY_REQUEST_NOTIFICATIONS_CREATE_OR_UPDATE],
+        )
+    ],
+)
+def create_or_update_privacy_request_notifications(
+    *, db: Session = Depends(deps.get_db), request_body: PrivacyRequestNotificationInfo
+) -> PrivacyRequestNotificationInfo:
+    """Create or update list of email addresses and number of failures for privacy request notifications."""
+    notification_info = {
+        "email": ", ".join(request_body.email_addresses),
+        "notify_after_failures": request_body.notify_after_failures,
+    }
+    info_check = PrivacyRequestNotifications.all(db)
+    if info_check:
+        info = info_check[0].update(db=db, data=notification_info)
+        return PrivacyRequestNotificationInfo(
+            email_addresses=info.email.split(", "),
+            notify_after_failures=info.notify_after_failures,
+        )
+
+    info = PrivacyRequestNotifications.create(db=db, data=notification_info)
+    return PrivacyRequestNotificationInfo(
+        email_addresses=info.email.split(", "),
+        notify_after_failures=info.notify_after_failures,
     )
 
 
@@ -715,7 +783,7 @@ async def resume_privacy_request(
     *,
     db: Session = Depends(deps.get_db),
     webhook: PolicyPreWebhook = Security(
-        verify_callback_oauth, scopes=[scopes.PRIVACY_REQUEST_CALLBACK_RESUME]
+        verify_callback_oauth, scopes=[PRIVACY_REQUEST_CALLBACK_RESUME]
     ),
     webhook_callback: PrivacyRequestResumeFormat,
 ) -> PrivacyRequestResponse: